Improved player movement and other minor details
Improved the player movement with proper acceleration.
Also changed the players collision box as it was hard to move through halls or small
areas without constantly bumping against the walls due to the shoulders.
Instead of having an entire large tiled area, I made two small rooms connected by a hall
which is enough and even better for testing.